refactor(cfd-math): Decompose gmg/mod.rs into ops and multigrid leaf modules
Part of ATLAS-ARCH-007.
